Added the 'query elapsed time' field to the VQL history
It will need some refactoring. The elapsed time emitted is the VariantModel member from the Variant View plugin. But due to a confusion I made between variants_loaded and variant_count_loaded (which are signals and slots that don't get triggered at the same time), we end up evaluating thee time for the previous query